Création de composants supplémentaires : - ArticleUpdate (fonctionnel, est masqué par défaut) - ArticleItem (fonctionnel) - Home-page (fonctionnel, page d'accueil de l'app) - Article-page (encore vide, page article, contiendra les commentaires)
